Set lang and prefix to null by default
So far I've been setting default values in the DB directly but I'm starting to think I'd much rather set those values in the application - this is a matter of separation of concerns.
Here the DB shouldn't have to know what the default prefix and the default language are, this makes the DB more efficient and also makes it more dynamic as you could in theory change the prefix and the language later.